Change type of GO_NEXT_SVG_GZ_SIZE to size_t (for Visual C++ compiler).
This hopefully fixes an error which occured on appveyor with MSVC.
Now the type is the same in src/StatusBar.h and the (generated)
build/wxm_toolbar_images.h file.

+++ src/StatusBar.h
@@ -36,7 +36,7 @@
 #include <memory>
 
 extern unsigned char GO_NEXT_SVG_GZ[];
-extern unsigned int  GO_NEXT_SVG_GZ_SIZE;
+extern size_t GO_NEXT_SVG_GZ_SIZE;
 
 #ifndef STATUSBAR_H
 #define STATUSBAR_H
